Refractored the Web3 Functions to One Module
