+++ package.json
@@ -10,11 +10,7 @@
     "@testing-library/user-event": "^12.1.10",
     "@truffle/hdwallet-provider": "1.0.40",
     "classnames": "^2.2.6",
-    "crypto-js": "^4.0.0",
     "file-saver": "^2.0.5",
-    "filepond": "^4.26.0",
-    "filepond-plugin-image-exif-orientation": "^1.0.11",
-    "filepond-plugin-image-preview": "^4.6.6",
     "ipfs-api": "^26.1.2",
     "ipfs-core": "^0.5.4",
     "js-string-compression": "^1.0.1",
@@ -24,9 +20,7 @@
     "node-sass": "^4.13.1",
     "react": "^16.8.0",
     "react-dom": "^16.8.0",
-    "react-dropzone": "^11.3.2",
     "react-file-drop": "^3.1.2",
-    "react-filepond": "^7.1.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "react-text-typist": "^1.0.5",
