--- src/Web3/AddFileHash/index.jsx
@@ -1,9 +1,0 @@
-const AddFile=async (contract,username,filehash,filename)=>{
-    const web3=window.web3;
-    const account = (await web3.eth.getAccounts())[0];
-    console.log(username,filehash,filename)
-    const result= await contract.methods.AddFileHash(username,filehash,filename).send({from : account});
-    return result;
-};
-
-export default AddFile;
