--- src/Web3/AddShareData/index.jsx
@@ -1,8 +1,0 @@
-const AddShareFile=async (contract,receiver,filehasharray)=>{
-    const web3=window.web3;
-    const account = (await web3.eth.getAccounts())[0];
-    const result= await contract.methods.AddShareFile(receiver,filehasharray).send({from : account});
-    return result;
-};
-
-export default AddShareFile;
