--- src/Web3/LoadWeb3/index.js
@@ -1,16 +1,0 @@
-import Web3 from "web3";
-
-const loadWeb3=async ()=> {
-    if (window.ethereum) {
-			window.web3 = new Web3(window.ethereum);
-			await window.ethereum.enable();
-		}
-		if (window.web3) {
-			window.web3 = new Web3(window.web3.currentProvider);
-		} else {
-			window.alert("please use metamask");
-		}
-
-  }
-
-export default loadWeb3;
